[None][chore] Wrap the swiglu into custom op to avoid redundant device copy.
